Add test-coverage and output, Restructure tests
- Add test-coverage:
  - Create `make report`: `report` returns the coverage of the
    respective tests by creating coverprofiles that are merged and
    outputted as html-file.

- Restructure tests:
  - Move integration tests to their respective driver-folders.
    Additionally the build-tag `+build integration` is set.
  - Minor adjustments that are necessary since the tests are moved to
    another directory.
