+++ examples/cgo/recorder/main_test.go
@@ -1,6 +1,7 @@
 // SPDX-FileCopyrightText: 2020 SAP SE
 //
 // SPDX-License-Identifier: Apache-2.0
+// +build integration
 
 package main
